corrected column id_telegram Integer to BigInteger

+++ Project_telegram/core/model/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import String, TIMESTAMP, Integer
+from sqlalchemy import String, TIMESTAMP, BigInteger
 
 from core.model.Base import BaseDB
 from core.model.id_mixin import IdPrKey
@@ -9,7 +9,7 @@ from core.model.id_mixin import IdPrKey
 
 class User(BaseDB, IdPrKey):
     name: Mapped[str] = mapped_column(String(50), default="Not name")
-    id_telegram: Mapped[int] = mapped_column(Integer, nullable=False)
+    id_telegram: Mapped[int] = mapped_column(BigInteger, nullable=False)
     created_at: Mapped[datetime] = mapped_column(
         TIMESTAMP(timezone=True),
         default=lambda: datetime.now(tz=timezone.utc),
